+++ example-nobundle/views/dashboard.js
@@ -46,7 +46,7 @@ export default {
         label: 'ID',
         key: 'id',
         render: (val, key, row) => '<a class="button" onclick="alert('+val+')">'+val+'<a>'
-        // can also fire off event
+        // can also fire off event - this.dispatchEvent(new CustomEvent('something', { detail: { val, row, key } }))
       },
       {
         label: 'Name',